CAMBIOS GENERALES. ESTACIONAMIENTOMANUAL REALIZADO pendiente x APP

+++ src/ar/edu/unq/po2/tpFinal/CompraEstacionamiento.java
@@ -5,7 +5,8 @@ import java.time.LocalDateTime;
 public class CompraEstacionamiento extends Compra {
 	private int cantHoras;
 
-	public CompraEstacionamiento(int ticketAsignado, PuntoDeVenta punto, LocalDateTime fechayHoraDeRecarga, int horas) {
+	public CompraEstacionamiento(double ticketAsignado, PuntoDeVenta punto, LocalDateTime fechayHoraDeRecarga,
+			int horas) {
 		this.setNroTicket(ticketAsignado);
 		this.setPuntoDeCompra(punto);
 		this.setFechaYHora(fechayHoraDeRecarga);
